Refactor Room component to use graphql module
